+++ index.es6.js
@@ -3,5 +3,18 @@ export { default as errors } from './source/errors'
 export { default as api } from './source/api'
 export { default as http } from './source/http'
 export { default as acl } from './source/acl'
-export { issue_jwt_token as jwt } from './source/middleware/authentication'
-export { generate_unique_filename as generate_unique_filename, generate_unique_filename as generateUniqueFilename } from './source/middleware/file upload'
+
+export
+{
+	issue_jwt_token as jwt,
+	verify_jwt_token as verify_jwt,
+	verify_jwt_token as verifyJwt
+}
+from './source/middleware/authentication'
+
+export
+{
+	generate_unique_filename,
+	generate_unique_filename as generateUniqueFilename
+}
+from './source/middleware/file upload'